fix: incorrect prometheus metrics
Due to the fact that a single test case in the configuration file actually contains multiple test cases (they are just grouped together), such as different combinations of threads and operations, each representing a distinct test scenario, we need to rename the test case names accordingly. This ensures that Prometheus can correctly collect and distinguish the test results without errors.

Signed-off-by: Liang Zheng <zhengliang0901@gmail.com>